Move Coordinator's & ViewController's to separate files
